Refactor/data : Enum type data constant dir로 이동
