+++ src/main/java/com/JEJU/yeonjigonji_backend/constant/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.JEJU.yeonjigonji_backend.entity;
+package com.JEJU.yeonjigonji_backend.constant;
 
 public enum OrderStatus {
     ORDER, CANCEL
